Add Houdini build scripts and improve Houdini support
Introduce Windows and PowerShell build scripts for building OpenDCC with Houdini's USD and dependencies.
Update CMake files to improve compatibility with Houdini builds, including conditional handling of Boost.Python, OSL, and MSVC preprocessor flags. Add new CMake Find modules for Imath and OpenEXR. Update documentation and .gitignore for Houdini builds. Add utility scripts for dependency checking and DLL management.

+++ src/packages/opendcc.ui.node_editor/lib/opendcc/ui/node_editor/layout.cpp
@@ -98,7 +98,7 @@ bool layout_items(const QVector<NodeItem*>& items, bool vertical)
 
     GVC_t* gvc = gvContext();
     char* dummy_dot;
-    uint32_t dummy_length;
+    size_t dummy_length;
 
     gvLayout(gvc, graph, "dot");
     auto res = gvRenderData(gvc, graph, "dot", &dummy_dot, &dummy_length);
